Add CallbackGroup
A CallbackGroup, by its name, groups callbacks for coming events. It is
an "abstract" class, it must be subclassed and added to
client.rooms.callback_groups for taking effect. Its defined methods will
be treated as callbacks as long as they have the correct signature
(Room, Event or Event[EventContent]). They will be called for the
respective upcoming event type.