Add CallbackGroup (#14)
* Add CallbackGroup

A CallbackGroup, by its name, groups callbacks for coming events. It is
an "abstract" class, it must be subclassed and added to
client.rooms.callback_groups to take effect. Its defined methods will
be treated as callbacks as long as they have the correct signature
(Room, Event or Event[EventContent]). They will be called for the
respective upcoming event type.

* Improve CallbackGroup tests

- Add async and state, and content tests
- Improve current TimelineEvent test

* Move CallbackGroup to rooms.py

Since CallbackGroup is not a big class, it is reasonable to move it to
rooms.py

* Shorten CallbackGroup code

Co-authored-by: miruka <miruka@disroot.org>